DittoConfig improvements
- Move DittoConfig to its own module (file).
- Add plumbing for cloud sync, persist. dir, and app token.
